[CSPM] Run commands as nobody by default
