+++ releasenotes/notes/cspm-command-check-user-145e687beab0af8e.yaml
@@ -1,0 +1,5 @@
+---
+enhancements:
+  - |
+    Run CSPM commands as a configurable user.
+    Defaults to 'nobody'.
